+++ packages/passport-local/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.13.0](https://github.com/koexjs/koex/compare/v0.12.4...v0.13.0) (2020-08-27)
+
+
+### Features
+
+* **passport:** refactor passport to add refresh token oauth mode support, and add passport.{getToken,setToken} ([9929d41](https://github.com/koexjs/koex/commit/9929d410275899147fcd8e2054e22d1b47b15136))
+
+
+
+
+
 ## 0.12.4 (2020-07-25)
 
 **Note:** Version bump only for package @koex/passport-local